% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.smoothrq}
\alias{coef.smoothrq}
\title{Extract smoothed quantile regression model coefficients}
\usage{
\method{coef}{smoothrq}(object, type = c("response", "smoothed"), ...)
}
\arguments{
\item{object}{an object for which the extraction of model coefficients is
    meaningful.}

\item{type}{The option \code{"response"} means that the smoothing will be inverted,
giving \eqn{\Theta H'}. This will result in a List containing one matrix for
each response. Alternatively, choosing \code{"smoothed"} will return \eqn{\Theta}
with the List containing one matrix for each degree.}

\item{...}{not used.}
}
\value{
a List of coefficient matrices.
}
\description{
Extract smoothed quantile regression model coefficients
}
\examples{
x <- matrix(rnorm(100 * 10), nrow = 100)
y <- matrix(rnorm(100 * 10), nrow = 100)
out <- smooth_rq(x, y, tau = c(.25, .5, .75))
cc1 <- coef(out)
cc2 <- coef(out, "smoothed")
}
