% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagmat.R
\name{lagmat}
\alias{lagmat}
\title{Create a matrix of leads and lags}
\usage{
lagmat(x, lags)
}
\arguments{
\item{x}{a univariate signal, observed regularly}

\item{lags}{a vector of positive (lags) and negative (leads) numbers}
}
\value{
a matrix
}
\description{
Create a matrix of leads and lags
}
\examples{
x <- 1:10
lagmat(x, c(-2, 0, 1, 2, 3))
}
