% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth-rq.R
\name{predict.smoothrq}
\alias{predict.smoothrq}
\title{Make predictions from a smoothed multi-output quantile regression}
\usage{
\method{predict}{smoothrq}(object, newdata, ...)
}
\arguments{
\item{object}{Object of class inheriting from \code{smooth_rq}}

\item{newdata}{A matrix or data frame in which to look for
variables with which to predict. May NOT be omitted.}

\item{...}{not used}
}
\value{
a List of predicted values, one matrix per response.
}
\description{
Make predictions from a smoothed multi-output quantile regression
}
\examples{
x <- matrix(rnorm(100 * 10), nrow = 100)
y <- matrix(rnorm(100 * 10), nrow = 100)
out <- smooth_rq(x, y, tau = c(.25, .5, .75))
p <- predict(out, newdata = x[1:10, ])
}
