% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth-rq.R
\name{smooth_rq}
\alias{smooth_rq}
\title{Smoothed multi-response quantile regression}
\usage{
smooth_rq(
  x,
  y,
  tau = 0.5,
  degree = 3L,
  intercept = TRUE,
  aheads = 1:n_models,
  ...
)
}
\arguments{
\item{x}{numeric matrix or data.frame of predictors. No intercept should be
included.}

\item{y}{numeric matrix or data.frame of response values}

\item{tau}{the quantile(s) to be estimated. This is a number (or vector of
values) strictly between 0 and 1.}

\item{degree}{the number of polynomials used for response smoothing. Must
be no more than the number of columns in y.}

\item{intercept}{logical. should the model be estimated with intercept}

\item{aheads}{Defaults to the vector \code{1:ncol(y)} but if the responses are
observed at a different spacing (or appear in a different order), that
information should be used here.}

\item{...}{additional arguments passed on to \link[quantreg:rq]{quantreg::rq}. Not all are
allowed.}
}
\value{
An object of class \code{smoothrq}
}
\description{
Smoothed multi-response quantile regression
}
\examples{
x <- matrix(rnorm(100 * 10), nrow = 100)
y <- matrix(rnorm(100 * 10), nrow = 100)
out <- smooth_rq(x, y, tau = c(.25, .5, .75))
}
